% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-utils.R
\name{alert}
\alias{alert}
\title{Alert}
\usage{
alert(..., class = "", color = "primary")
}
\arguments{
\item{...}{Content of the alert.}

\item{class}{Any additional class.}

\item{color}{Color of the alert.}
}
\description{
Create a bootstrap 4 alert.
}
\keyword{internal}
