% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_app_navbar.R
\name{app_nav_server}
\alias{app_nav_server}
\title{Server Logic for the Application Navbar}
\usage{
app_nav_server(input, output, session)
}
\arguments{
\item{input}{Shiny's input object.}

\item{output}{Shiny's output object.}

\item{session}{Shiny's session object.}
}
\value{
None.
}
\description{
This function provides the server-side logic for the Shiny application
using the navbar UI created by \code{app_nav_ui}. It is a placeholder function
with no implemented logic.
}
