% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_app_navbar.R
\name{app_nav_ui}
\alias{app_nav_ui}
\title{Create the User Interface for the Application Navbar}
\usage{
app_nav_ui(theme = "light")
}
\arguments{
\item{theme}{\code{"dark"} or \code{"light"}}
}
\value{
A UI definition that can be passed to a Shiny app's \code{ui} parameter.
}
\description{
This function generates the UI for a Shiny application that uses a Bootstrap 5 navbar.
The navbar contains multiple panels, each with a sidebar layout or card-based tabs.
}
