% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{app_sys}
\alias{app_sys}
\title{Access files in the current app}
\usage{
app_sys(...)
}
\arguments{
\item{...}{character vectors, specifying subdirectory and file(s)
within your package. The default, none, returns the root of the app.}
}
\description{
NOTE: If you manually change your package name in the DESCRIPTION,
don't forget to change it here too, and in the config file.
For a safer name change mechanism, use the \code{golem::set_golem_name()} function.
}
