% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-utils.R
\name{badge}
\alias{badge}
\title{Badge}
\usage{
badge(..., class = "", color = "dark", pill = FALSE, .tag = shiny::span)
}
\arguments{
\item{...}{Content of the badge.}

\item{class}{Any additional class.}

\item{color}{Color of the badge.}

\item{pill}{Whether to create a pill badge.}

\item{.tag}{Tag to use (\code{span}, or \code{a})}
}
\description{
Add a badge.
}
\keyword{internal}
