% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthstone.R
\name{birthstoneServer}
\alias{birthstoneServer}
\title{Month birthstone (server)}
\usage{
birthstoneServer(id, month)
}
\arguments{
\item{id}{module ID}

\item{month}{month input}
}
\value{
\code{month} and \code{stone} values from \code{stones} data.
}
\description{
Month birthstone (server)
}
