% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthstone.R
\name{birthstoneUI}
\alias{birthstoneUI}
\title{Month birthstone (UI)}
\usage{
birthstoneUI(id)
}
\arguments{
\item{id}{module ID}
}
\value{
birthstone UI module elements
}
\description{
Month birthstone (UI)
}
