% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttonServer.R
\name{buttonServer}
\alias{buttonServer}
\title{Button Server Function}
\usage{
buttonServer(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\value{
None
}
\description{
This function defines the server logic of the Shiny app.
}
\details{
The main server logic includes the customer button server module to capture
user input and the greeting server module to display the personalized
greeting message. It also includes an output to display reactive values for
debugging purposes.
}
