% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttonUI.R
\name{buttonUI}
\alias{buttonUI}
\title{Button UI Function}
\usage{
buttonUI()
}
\value{
A tagList containing the main UI layout.
}
\description{
This function creates the main UI layout of the Shiny app.
}
\details{
The main UI layout includes the customer button module for user input and
the greeting module for displaying the greeting message. It also includes an
output to display reactive values for debugging purposes.
}
