% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_installed.R
\name{check_installed}
\alias{check_installed}
\title{Check if package is installed}
\usage{
check_installed(package)
}
\arguments{
\item{package}{string, name of package}
}
\value{
invisible
}
\description{
An example function for demonstrating how to use \code{testthat}'s
mocking functions.
}
\examples{
check_installed("foo")
check_installed("base")
}
