% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{col_12}
\alias{col_12}
\title{Column wrapper}
\usage{
col_12(...)
}
\arguments{
\item{...}{arguments from \code{shiny::column()}}
}
\description{
These are convenient wrappers around
\code{column(12, ...)}
}
