% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot.R
\name{compare_plot}
\alias{compare_plot}
\title{Create comparison chart (report)}
\usage{
compare_plot(data, x, y, color = NULL, alpha = 0.7, size = 3, title = NULL)
}
\arguments{
\item{data}{movies dataset}

\item{x}{x variable from UI}

\item{y}{y variable from UI}

\item{color}{color variable from UI}

\item{alpha}{alpha aesthetic from UI}

\item{size}{size aesthetic from UI}

\item{title}{title text from UI}
}
\value{
ggplot2 object
}
\description{
Create comparison chart (report)
}
