% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bar_plot.R
\name{count_bar_plot}
\alias{count_bar_plot}
\title{Create a Horizontal Bar Plot of Movies Count}
\usage{
count_bar_plot(data, chr_var, x_lab)
}
\arguments{
\item{data}{A dataframe containing the filtered movies data}

\item{chr_var}{The categorical variable to count (as a symbol)}

\item{x_lab}{Label for the x-axis (title of the categorical variable)}
}
\value{
A ggplot2 object displaying a horizontal bar plot
}
\description{
This function generates a horizontal bar plot showing the frequency counts of a
selected categorical variable from the movies dataset, filtered by year range.
}
\details{
The function creates a bar chart with the categorical variable on the y-axis
(flipped coordinates) and the count on the x-axis. Categories are ordered by frequency
(most frequent at the top).
}
\examples{
# create plot
count_bar_plot(data = movies, chr_var = "mpaa_rating", x_lab = "mpaa_rating")

}
