% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_template_dir}
\alias{create_template_dir}
\title{Create report directory if it doesn't exist}
\usage{
create_template_dir()
}
\value{
directory
}
\description{
Create report directory if it doesn't exist
}
