% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{data1}
\alias{data1}
\title{data1}
\usage{
data1
}
\description{
The data set is used in Sample 16.1.
The data set consists of 2 variables and 10 observations.
This data is imaginary.
}
\keyword{datasets}
