% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_app_server.R
\name{db_app_server}
\alias{db_app_server}
\title{Define the Server Logic for the Shiny Application}
\usage{
db_app_server(input, output, session, .dev = FALSE)
}
\arguments{
\item{input, output, session}{Standard Shiny server function arguments.}
}
\value{
A \code{shiny.server} function handling user inputs and outputs.
}
\description{
This function sets up server-side logic, including reactive movie data,
input modules, and filtering operations. It initializes reactive expressions
for user input and manages interactions between different UI components.
}
