% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_app_ui.R
\name{db_app_ui}
\alias{db_app_ui}
\title{Define the User Interface for the Shiny Application}
\usage{
db_app_ui()
}
\value{
A \code{shiny.tag.list} object representing the application UI.
}
\description{
This function constructs the UI layout of the Shiny application, including
input filters, slider controls, and movie selection elements.
}
