% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_con.R
\name{db_con}
\alias{db_con}
\title{Establish a Database Connection}
\usage{
db_con(pkg, db)
}
\arguments{
\item{pkg}{A character string specifying the name of the R package where
the database file is located.}

\item{db}{A character string specifying the name of the SQLite database
file (e.g., \code{"movies.db"}).}
}
\value{
A \code{DBIConnection} object representing the connection to the database.
}
\description{
This function connects to an SQLite database within an R package's
\code{extdata} directory.
}
