% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint-utils.R
\name{default_json_serialiser}
\alias{default_json_serialiser}
\title{Default Serialiser}
\usage{
default_json_serialiser(data)
}
\arguments{
\item{data}{Dataset to serialise.}
}
\description{
Default serialiser used by leprechaun.
To change this serialiser set the
\code{LEPRECHAUN_SERLIASER} option to the function
you want to use. This function must accept a
single argument, the dataset to serialise.
}
\keyword{internal}
