% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_app_navbar.R
\name{demo_app_navbar}
\alias{demo_app_navbar}
\title{Run the Navbar Demo Application}
\usage{
demo_app_navbar(...)
}
\arguments{
\item{...}{arguments passed to \code{app_nav_ui()}}
}
\value{
A Shiny app object.
}
\description{
This function launches a Shiny application that demonstrates the UI and
server logic created by \code{app_nav_ui} and \code{app_nav_server}.
}
