% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_type.R
\name{display_type}
\alias{display_type}
\title{Shiny app display mode helper}
\usage{
display_type(run = "w")
}
\arguments{
\item{run}{where to launch app:
\itemize{
\item \code{"p"} = launch in viewer pane
\item \code{"b"} = launch in external browser
\item \code{"w"} = launch in window (default)
}}
}
\value{
notification of \code{shinyViewerType} option
}
\description{
Shiny app display mode helper
}
\section{IDE detection}{

The function will check if the application is being run in RStudio
(with \code{Sys.getenv()}) and only apply the \code{shiny.launch.browser}
option if \code{TRUE}.
}

