% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-js.R
\name{enable}
\alias{enable}
\alias{disable}
\title{Enable and Disable}
\usage{
disable(
  selector,
  process_selector = TRUE,
  session = shiny::getDefaultReactiveDomain()
)

enable(
  selector,
  process_selector = TRUE,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{selector}{A CSS selector. If no selector
is recognised and \code{process_selector} is set to
\code{TRUE} then it is assumed to be an id.
e.g.: \code{show("hello")} will reveal the element with
\code{id} or \code{inputId} of \code{hello}.}

\item{process_selector}{Whether to pre-process the
\code{selector} argument to ensure a valid CSS selector
is passed.}
}
\description{
Enable and disable inputs.
}
\keyword{internal}
