% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_award_wins.R
\name{filter_award_wins}
\alias{filter_award_wins}
\title{Filter Movie Awards (support function)}
\usage{
filter_award_wins(award_data, award, start_yr, end_yr)
}
\arguments{
\item{award_data}{data to filter}

\item{award}{character string}

\item{start_year}{integer}

\item{end_year}{integer}
}
\value{
A data frame containing movies that won the specified award within
the given year range.
}
\description{
This function filters movie awards data.
}
