% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_award_recipients.R
\name{get_award_recipients}
\alias{get_award_recipients}
\title{Get award recipients from the movies dataset}
\usage{
get_award_recipients(data, award, year)
}
\arguments{
\item{data}{A data frame containing the \code{movies} dataset.}

\item{award}{A character string, either \code{"picture/director"} or
\code{"actor/actress"}.}

\item{year}{Integer, the year a range is calculated from. If
\code{"picture/director"}, +/- 2 years. If \code{"actor/actress"}, +/- 1 year.}
}
\value{
A \code{tibble} with selected columns depending on the award type:
\itemize{
\item \code{award == "picture/director"}: returns \code{title}, \code{release_year}, \code{director}
\item \code{award == "actor/actress"}: returns \code{title}, \code{release_year}, \code{star}
}
}
\description{
This function filters the \code{movies} dataset based on the selected award type
("picture/director" or "actor/actress") and a specified year.
}
