% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint-utils.R
\name{get_json_serialiser}
\alias{get_json_serialiser}
\title{Retrieve the Default Serialiser}
\usage{
get_json_serialiser()
}
\description{
Retrieves the \code{LEPRECHAUN_SERIALISER} option,
if not defined, it returns \code{default_json_serialiser}.
}
\keyword{internal}
