% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_movies_app.R
\name{ggp2_movies_app}
\alias{ggp2_movies_app}
\title{Development \code{ggplot2movies} app standalone function}
\usage{
ggp2_movies_app(test = FALSE, run = "w")
}
\arguments{
\item{test}{logical, run in \code{test.mode}? Defaults to \code{FALSE}.}

\item{run}{where to launch app:
\itemize{
\item \code{p} = launch in viewer pane
\item \code{b} = launch in external browser
\item \code{w} = launch in window
\item \code{NULL} = \code{getOption(x = 'shiny.launch.browser')}
}}
}
\value{
shiny app
}
\description{
Wrapper function for \code{shinyAppDir()}
}
