% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint-utils.R
\name{http_response_json}
\alias{http_response_json}
\title{Return a JSON Response}
\usage{
http_response_json(content, headers = list())
}
\description{
Return a json response, used for \code{session$registerDataObj}.
}
\keyword{internal}
