% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{if_not_NA}
\alias{if_not_NA}
\alias{\%|NA|\%}
\title{If x is \code{NA}, return y, otherwise return x}
\usage{
x \%|NA|\% y
}
\arguments{
\item{x, y}{Two elements to test, one potentially \code{NA}}
}
\description{
If x is \code{NA}, return y, otherwise return x
}
\examples{
NA \%|NA|\% 1
}
