% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{if_not_NULL}
\alias{if_not_NULL}
\alias{\%||\%}
\title{If x is \code{NULL}, return y, otherwise return x}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{Two elements to test, one potentially \code{NULL}}
}
\description{
If x is \code{NULL}, return y, otherwise return x
}
\examples{
NULL \%||\% 1
}
