% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{jq_hide}
\alias{jq_hide}
\title{Hide an elements by calling jquery hide on it}
\usage{
jq_hide(id)
}
\arguments{
\item{id}{the id of the element to hide}
}
\description{
Hide an elements by calling jquery hide on it
}
