% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_formatter.R
\name{lab_formatter}
\alias{lab_formatter}
\title{Convert snake_case to Sentence Case}
\usage{
lab_formatter(x)
}
\arguments{
\item{x}{A character string in \code{snake_case}.}
}
\value{
A character string converted to \verb{Sentence Case}.
}
\description{
\code{lab_formatter()} converts a given \code{snake_case} string to a \verb{Sentence Case} string.
}
\details{
The \code{lab_formatter} function splits the input string by underscores, capitalizes
the first letter of each word, and combines the words with spaces.
This is particularly useful for converting variable names into more readable
labels for plots.

For example, when used with the \code{ggplot2::labs()} function, \code{lab_formatter} can
automatically convert \code{snake_case} labels into \verb{Sentence Case} labels for
better readability.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(ggplot2)

# Example data
df <- data.frame(snake_case_var = rnorm(10))

# Example usage in ggplot2
ggplot(df, aes(x = snake_case_var)) +
  geom_histogram() +
  labs(x = lab_formatter("snake_case_var"))
}\if{html}{\out{</div>}}

This will result in an x-axis label "Snake Case Var".
}
\examples{
lab_formatter("snake_case_example")
lab_formatter("space example")

}
