% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_debug.R
\name{launch_debug}
\alias{launch_debug}
\title{Launch a Debugging Shiny App}
\usage{
launch_debug()
}
\value{
Launches the Shiny app in the default web browser.
}
\description{
This function launches a Shiny app that demonstrates the use of
\code{observe()}, \code{isolate()}, \code{reactive()}, and \code{bindEvent()} functions
in a simple interactive setting. The app allows users to see the behavior
of these reactive constructs in action, particularly how \code{observe()} can be
used with \code{isolate()} to control reactive dependencies and how \code{bindEvent()}
triggers specific actions.
}
\details{
The app consists of a sidebar with an action button and a numeric input
field. The main panel displays two outputs: one showing the isolated value
of the numeric input when the button is clicked, and another displaying the
reactive value that updates automatically as the input changes. Key reactive
functions used in the app include:
\itemize{
\item \code{reactive()}: Creates a reactive expression that updates whenever the
numeric input changes.
\item \code{observe()}: Monitors events and performs actions, here used with
\code{bindEvent()} to execute code when the button is clicked.
\item \code{isolate()}: Prevents reactive dependencies within \code{observe()}, ensuring
the value is only updated when the button is pressed, not when the input
changes directly.
\item \code{bindEvent()}: Binds an event (button click) to the execution of a specific
code block.
}
}
\examples{
if (interactive()) {
  launch_debug()
}

}
