% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-utils.R
\name{leprechaun-columns}
\alias{leprechaun-columns}
\alias{col1}
\alias{col2}
\alias{col3}
\alias{col4}
\alias{col5}
\alias{col6}
\alias{col7}
\alias{col8}
\alias{col9}
\alias{col10}
\alias{col11}
\alias{col12}
\title{Columns}
\usage{
col1(...)

col2(...)

col3(...)

col4(...)

col5(...)

col6(...)

col7(...)

col8(...)

col9(...)

col10(...)

col11(...)

col12(...)
}
\arguments{
\item{...}{Any valid htmltools content and tags
to pass to \code{column}.}
}
\description{
Shortcuts to create columns of specific width.
}
\keyword{internal}
