% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{list_to_p}
\alias{list_to_p}
\title{Turn an R list into corresponding HTML paragraph tags}
\usage{
list_to_p(list, class = NULL)
}
\arguments{
\item{list}{an R list}

\item{class}{a class for the paragraph tags}
}
\value{
An HTML tag
}
\description{
Turn an R list into corresponding HTML paragraph tags
}
\examples{
list_to_p(c("This is the first paragraph", "this is the second paragraph"))

}
