% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{make_action_button}
\alias{make_action_button}
\title{Make the current tag behave like an action button}
\usage{
make_action_button(tag, inputId = NULL)
}
\arguments{
\item{tag}{Any compatible tag.}

\item{inputId}{Unique id. This will host the input value to be used
on the server side.}
}
\value{
The modified tag with an extra id and the action button class.
}
\description{
Only works with compatible tags like button or links
}
\examples{
if (interactive()) {
  library(shiny)

  link <- a(href = "#", "My super link", style = "color: lightblue;")

  ui <- fluidPage(
    make_action_button(link, inputId = "mylink")
  )

  server <- function(input, output, session) {
    observeEvent(input$mylink, {
      showNotification("Pouic!")
    })
  }

  shinyApp(ui, server)
}
}
