% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{make_changed_rVals_inputs}
\alias{make_changed_rVals_inputs}
\title{Make changed \code{reactiveValues()} for module test}
\usage{
make_changed_rVals_inputs()
}
\value{
reactive test values
}
\description{
Used for testing \code{\link[=mod_scatter_display_server]{mod_scatter_display_server()}}
}
\examples{
require(shiny)
rVals <- make_changed_rVals_inputs()
isolate(rVals$inputs())
}
