% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_aes_input.R
\name{mod_aes_input_ui}
\alias{mod_aes_input_ui}
\title{Aesthetics Input Module - UI}
\usage{
mod_aes_input_ui(id)
}
\arguments{
\item{id}{\emph{(character)} Namespace ID for the module.}
}
\value{
A \code{shiny::tagList()} containing UI elements for variable
selection.
}
\description{
Creates a UI for selecting attributes for a scatter plot.
}
\section{Details}{

\code{mod_aes_input_ui()} provides UI controls for customizing scatter plots.
It includes:
\itemize{
\item \strong{Sliders}:
\itemize{
\item Transparency (\code{alpha}) ranging from 0 to 1.
\item Point size ranging from 0 to 5.
}
\item \strong{Text Input}:
\itemize{
\item Plot title.
}
}

The function uses \code{shiny::NS()} to namespace all input IDs.
}

\seealso{
\itemize{
\item \code{\link[=mod_aes_input_server]{mod_aes_input_server()}}  for server-side logic.
\item \href{https://shiny.rstudio.com/reference/shiny/latest/tagList.html}{\code{shiny::tagList()}}
for containing UI elements.
\item \href{https://shiny.rstudio.com/reference/shiny/latest/NS.html}{\code{shiny::NS()}}
for namespacing.
}
}
\concept{\strong{Aesthetics Input Module}}
