% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_awards_point.R
\name{mod_awards_point_server}
\alias{mod_awards_point_server}
\title{Server Logic for Scatter Plot Text Repel}
\usage{
mod_awards_point_server(id, vals)
}
\arguments{
\item{id}{Shiny module ID.}

\item{vals}{Reactive list of plot parameters.}
}
\value{
Renders a scatter plot.
}
\description{
Server Logic for Scatter Plot Text Repel
}
