% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_awards_point.R
\name{mod_awards_point_ui}
\alias{mod_awards_point_ui}
\title{User Interface for Scatter Plot Text Repel}
\usage{
mod_awards_point_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for displaying a scatter plot.
}
\description{
User Interface for Scatter Plot Text Repel
}
