% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boxplot.R
\name{mod_boxplot_server}
\alias{mod_boxplot_server}
\title{Server Logic for Box Plot}
\usage{
mod_boxplot_server(id, vals)
}
\arguments{
\item{id}{Shiny module ID.}

\item{vals}{Reactive list of plot parameters.}
}
\value{
Renders a box plot based on selected parameters.
}
\description{
Server Logic for Box Plot
}
\section{Reactive Inputs}{

\itemize{
\item \code{vals}: Reactive list of plot parameters.
}
}

\seealso{
\code{\link[=mod_boxplot_ui]{mod_boxplot_ui()}}
}
