% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compare_point.R
\name{mod_compare_point_server}
\alias{mod_compare_point_server}
\title{Server Logic for Scatter Plot}
\usage{
mod_compare_point_server(id, vals)
}
\arguments{
\item{id}{Shiny module ID.}

\item{vals}{Reactive list of plot parameters.}
}
\value{
Renders a scatter plot based on selected parameters.
}
\description{
Server Logic for Scatter Plot
}
\details{
This server function renders a scatter plot using plotly based on
user-selected variables. It includes logging at different levels to track
execution and potential errors.
}
\seealso{
\code{\link[=mod_compare_point_ui]{mod_compare_point_ui()}}
}
