% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compare_vars.R
\name{mod_compare_vars_ui}
\alias{mod_compare_vars_ui}
\title{UI module for comparing movie variables}
\usage{
mod_compare_vars_ui(id)
}
\arguments{
\item{id}{A string. The module ID.}
}
\value{
A UI element containing inputs for selecting variables to compare
in a scatter plot, including X and Y variables, color variable, and
customization controls for alpha, size, and plot title.
}
\description{
UI module for comparing movie variables
}
