% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_count_vars.R
\name{mod_count_vars_ui}
\alias{mod_count_vars_ui}
\title{User Interface for Count Tab Input}
\usage{
mod_count_vars_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for selecting a numeric input.
}
\description{
User Interface for Count Tab Input
}
\section{Inputs}{

\itemize{
\item \code{chr_var}: Selected grouping variable
\item \code{start_year}: Selected 'Theater Release' start year.
\item \code{end_year}: Selected 'Theater Release' end year.
}
}

\seealso{
\code{\link[=mod_count_vars_server]{mod_count_vars_server()}}
}
