% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_counts_bar.R
\name{mod_counts_bar_server}
\alias{mod_counts_bar_server}
\title{Server Logic for Horizontal Bar Plot}
\usage{
mod_counts_bar_server(id, vals)
}
\arguments{
\item{id}{Shiny module ID.}

\item{vals}{Reactive list of plot parameters.}
}
\value{
Renders a horizontal bar plot based on selected parameters.
}
\description{
Server Logic for Horizontal Bar Plot
}
