% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_counts_bar.R
\name{mod_counts_bar_ui}
\alias{mod_counts_bar_ui}
\title{User Interface for Horizontal Bar Plot}
\usage{
mod_counts_bar_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for displaying a horizontal bar plot.
}
\description{
User Interface for Horizontal Bar Plot
}
