% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_counts_tbl.R
\name{mod_counts_tbl_ui}
\alias{mod_counts_tbl_ui}
\title{User Interface for Counts Data Table}
\usage{
mod_counts_tbl_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for displaying a data table.
}
\description{
User Interface for Counts Data Table
}
