% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_counts_txt.R
\name{mod_counts_txt_server}
\alias{mod_counts_txt_server}
\title{Server Logic for Counts Text}
\usage{
mod_counts_txt_server(id, vals)
}
\arguments{
\item{id}{Shiny module ID.}

\item{vals}{Reactive list of table parameters.}
}
\value{
Renders a data table based on selected parameters.
}
\description{
Server Logic for Counts Text
}
\section{Reactive Inputs}{

\itemize{
\item \code{vals}: Reactive list of table parameters.
}
}

\seealso{
\code{\link[=mod_counts_txt_ui]{mod_counts_txt_ui()}}
}
