% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_counts_vars.R
\name{mod_counts_vars_ui}
\alias{mod_counts_vars_ui}
\title{User Interface for Count Tab Input}
\usage{
mod_counts_vars_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for selecting a numeric input.
}
\description{
User Interface for Count Tab Input
}
