% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_counts_vbox.R
\name{mod_counts_vbox_ui}
\alias{mod_counts_vbox_ui}
\title{User Interface for Counts Text}
\usage{
mod_counts_vbox_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for displaying a the time span.
}
\description{
User Interface for Counts Text
}
