% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_counts_waffle.R
\name{mod_counts_waffle_ui}
\alias{mod_counts_waffle_ui}
\title{User Interface for Vertical Bar Plot}
\usage{
mod_counts_waffle_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for displaying a vertical bar plot.
}
\description{
User Interface for Vertical Bar Plot
}
