% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cust_button.R
\name{mod_cust_button_server}
\alias{mod_cust_button_server}
\title{Customer Button Server Module}
\usage{
mod_cust_button_server(id)
}
\arguments{
\item{id}{A unique identifier for the module.}
}
\value{
A reactive expression containing the user input from the text field.
}
\description{
This server module captures the user input from the text field and makes it
available as a reactive expression.
}
\details{
This server module works in conjunction with the \code{mod_cust_button_ui}
function to capture user input. The reactive expression returned by this
module can be used by other modules to access the input value.
}
