% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cust_button.R
\name{mod_cust_button_ui}
\alias{mod_cust_button_ui}
\title{Customer Button UI Module}
\usage{
mod_cust_button_ui(id)
}
\arguments{
\item{id}{A unique identifier for the module.}
}
\value{
A tagList containing the UI elements for the customer button module.
}
\description{
This UI module creates a simple form with a text input for the user to enter
their name and a submit button.
}
\details{
This UI module works in conjunction with the \code{mod_cust_button_server}
function to handle user input. The input value is captured and can be used
in other server modules.
}
