% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_table.R
\name{mod_data_table_server}
\alias{mod_data_table_server}
\title{Data Table Module Server}
\usage{
mod_data_table_server(id, data, inputs)
}
\arguments{
\item{id}{Module namespace}

\item{data}{Reactive data}

\item{inputs}{List of reactive inputs from inputs module}
}
\value{
Reactive filtered data
}
\description{
Data Table Module Server
}
