% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dist_box.R
\name{mod_dist_box_ui}
\alias{mod_dist_box_ui}
\title{User Interface for Box Plots}
\usage{
mod_dist_box_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for displaying a box plot.
}
\description{
User Interface for Box Plots
}
