% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dist_raincloud.R
\name{mod_dist_raincloud_ui}
\alias{mod_dist_raincloud_ui}
\title{User Interface for Raincloud Plots}
\usage{
mod_dist_raincloud_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for displaying a rain cloud plot.
}
\description{
User Interface for Raincloud Plots
}
