% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_report.R
\name{mod_download_report_server}
\alias{mod_download_report_server}
\title{Download Report Module Server}
\usage{
mod_download_report_server(id, plot_data, table_data, inputs)
}
\arguments{
\item{id}{Module namespace}

\item{plot_data}{List containing plot object}

\item{table_data}{Reactive filtered data}

\item{inputs}{List of reactive inputs from inputs module}
}
\value{
None
}
\description{
Download Report Module Server
}
