% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_report.R
\name{mod_download_report_ui}
\alias{mod_download_report_ui}
\title{Download Report Module UI}
\usage{
mod_download_report_ui(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
Shiny UI element
}
\description{
Download Report Module UI
}
