% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filters.R
\name{mod_filters_server}
\alias{mod_filters_server}
\title{Reactable table shiny module (server)}
\usage{
mod_filters_server(id, movies, num, var, txt)
}
\arguments{
\item{id}{shiny module id}

\item{movies}{movies connection}

\item{num}{slider inputs}

\item{var}{variable inputs}

\item{txt}{text inputs}
}
\description{
Reactable table shiny module (server)
}
