% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_greeting.R
\name{mod_greeting_server}
\alias{mod_greeting_server}
\title{Greeting Server Module}
\usage{
mod_greeting_server(id, txt)
}
\arguments{
\item{id}{A unique identifier for the module.}

\item{txt}{A reactive expression containing the user input.}
}
\value{
None
}
\description{
This server module observes the user input and updates the greeting message
accordingly.
}
\details{
This server module works in conjunction with the \code{mod_greeting_ui} function
to display a personalized greeting message. It observes the reactive
expression from \code{mod_cust_button_server} and updates the text output
accordingly.
}
