% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_greeting.R
\name{mod_greeting_ui}
\alias{mod_greeting_ui}
\title{Greeting UI Module}
\usage{
mod_greeting_ui(id)
}
\arguments{
\item{id}{A unique identifier for the module.}
}
\value{
A tagList containing the UI elements for displaying the greeting
message.
}
\description{
This UI module creates a text output to display a greeting message.
}
\details{
This UI module works in conjunction with the \code{mod_greeting_server} function
to display a personalized greeting message based on user input.
}
