% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hbar.R
\name{mod_hbar_ui}
\alias{mod_hbar_ui}
\title{User Interface for Horizontal Bar Plot}
\usage{
mod_hbar_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for displaying a horizontal bar plot.
}
\description{
User Interface for Horizontal Bar Plot
}
\section{Outputs}{

\itemize{
\item \code{hbar_graph}: Horizontal bar plot.
}
}

\seealso{
\code{\link[=mod_hbar_server]{mod_hbar_server()}}
}
