% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_inputs.R
\name{mod_inputs_server}
\alias{mod_inputs_server}
\title{Inputs Module Server}
\usage{
mod_inputs_server(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
List of reactive values
}
\description{
Inputs Module Server
}
