% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_num_input.R
\name{mod_num_input_server}
\alias{mod_num_input_server}
\title{Server Logic for Numeric Input}
\usage{
mod_num_input_server(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A reactive list containing the selected year.
}
\description{
Server Logic for Numeric Input
}
\section{Reactive Values}{

\itemize{
\item \code{start_year}: Selected start theater release year.
\item \code{end_year}: Selected end theater release year.
}
}

\seealso{
\code{\link[=mod_num_input_ui]{mod_num_input_ui()}}
}
