% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_num_input.R
\name{mod_num_input_ui}
\alias{mod_num_input_ui}
\title{User Interface for Numeric Input}
\usage{
mod_num_input_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for selecting a numeric input.
}
\description{
User Interface for Numeric Input
}
\section{Inputs}{

\itemize{
\item \code{year}: Theater release year.
}
}

\seealso{
\code{\link[=mod_num_input_server]{mod_num_input_server()}}
}
