% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot.R
\name{mod_plot_server}
\alias{mod_plot_server}
\title{plot Server Functions}
\usage{
mod_plot_server(id, var_inputs)
}
\arguments{
\item{id}{module namespace id}

\item{var_inputs}{inputs from mod_var}
}
\value{
shiny server module
}
\description{
plot Server Functions
}
