% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_point_text.R
\name{mod_point_text_server}
\alias{mod_point_text_server}
\title{Server Logic for Scatter Plot Text Repel}
\usage{
mod_point_text_server(id, vals)
}
\arguments{
\item{id}{Shiny module ID.}

\item{vals}{Reactive list of plot parameters.}
}
\value{
Renders a scatter plot based on selected parameters.
}
\description{
Server Logic for Scatter Plot Text Repel
}
\section{Reactive Inputs}{

\itemize{
\item \code{vals}: Reactive list of plot parameters.
}
}

\seealso{
\code{\link[=mod_point_text_ui]{mod_point_text_ui()}}, \code{\link[=mod_awards_vars_server]{mod_awards_vars_server()}}, \code{\link[=mod_awards_vars_ui]{mod_awards_vars_ui()}}
}
