% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_point_vars.R
\name{mod_point_vars_server}
\alias{mod_point_vars_server}
\title{Server Logic for Variable Selection}
\usage{
mod_point_vars_server(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A reactive list of selected variables and plot attributes.
}
\description{
Server Logic for Variable Selection
}
\section{Reactive Values}{

\itemize{
\item \code{x}: X variable.
\item \code{y}: Y variable.
\item \code{c}: Color variable.
\item \code{a}: Alpha transparency.
\item \code{s}: Point size.
\item \code{t}: Plot title.
}
}

\seealso{
\code{\link[=mod_point_vars_ui]{mod_point_vars_ui()}}
}
