% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_points_plotly.R
\name{mod_points_plotly_server}
\alias{mod_points_plotly_server}
\title{Scatter-plot (plotly) shiny module (server)}
\usage{
mod_points_plotly_server(id, data, vars)
}
\arguments{
\item{id}{shiny module id}
}
\description{
Scatter-plot (plotly) shiny module (server)
}
