% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_points_plotly.R
\name{mod_points_plotly_ui}
\alias{mod_points_plotly_ui}
\title{Scatter-plot (plotly) shiny module (UI)}
\usage{
mod_points_plotly_ui(id)
}
\arguments{
\item{id}{shiny module id}
}
\description{
Scatter-plot (plotly) shiny module (UI)
}
