% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rainclouds.R
\name{mod_raincloud_ui}
\alias{mod_raincloud_ui}
\title{User Interface for Raincloud Plots}
\usage{
mod_raincloud_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for displaying a rain cloud plot.
}
\description{
User Interface for Raincloud Plots
}
\section{Outputs}{

\itemize{
\item \code{genre}: Vertical bar plot.
}
}

\seealso{
\code{\link[=mod_raincloud_server]{mod_raincloud_server()}}
}
