% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report_download.R
\name{mod_report_download_server}
\alias{mod_report_download_server}
\title{Download R Markdown Report (server)}
\usage{
mod_report_download_server(
  id,
  compare_values,
  count_values,
  dist_values,
  award_values
)
}
\arguments{
\item{id}{}
}
\value{
download button for report
}
\description{
Download R Markdown Report (server)
}
