% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_report.R
\name{mod_report_download_ui}
\alias{mod_report_download_ui}
\title{Download report module (UI)}
\usage{
mod_report_download_ui(id)
}
\arguments{
\item{id}{module id}
}
\value{
module UI
}
\description{
Download report module (UI)
}
