% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ridgeline.R
\name{mod_ridgeline_ui}
\alias{mod_ridgeline_ui}
\title{User Interface for Ridgeline Plots}
\usage{
mod_ridgeline_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for displaying a rain cloud plot.
}
\description{
User Interface for Ridgeline Plots
}
\section{Outputs}{

\itemize{
\item \code{genre}: Vertical bar plot.
}
}

\seealso{
\code{\link[=mod_ridgeline_server]{mod_ridgeline_server()}}
}
