% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_inputs.R
\name{mod_select_input_server}
\alias{mod_select_input_server}
\title{select inputs shiny module (server)}
\usage{
mod_select_input_server(id)
}
\arguments{
\item{id}{shiny module id}
}
\description{
select inputs shiny module (server)
}
