% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_inputs.R
\name{mod_select_input_ui}
\alias{mod_select_input_ui}
\title{select inputs shiny module (UI)}
\usage{
mod_select_input_ui(id)
}
\arguments{
\item{id}{shiny module id}
}
\description{
select inputs shiny module (UI)
}
