% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_slider_inputs.R
\name{mod_slider_input_ui}
\alias{mod_slider_input_ui}
\title{slider inputs shiny module (UI)}
\usage{
mod_slider_input_ui(id)
}
\arguments{
\item{id}{shiny module id}
}
\description{
slider inputs shiny module (UI)
}
