% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table_display.R
\name{mod_table_display_ui}
\alias{mod_table_display_ui}
\title{Reactable table shiny module (UI)}
\usage{
mod_table_display_ui(id)
}
\arguments{
\item{id}{shiny module id}
}
\description{
Reactable table shiny module (UI)
}
