% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tbl.R
\name{mod_tbl_ui}
\alias{mod_tbl_ui}
\title{User Interface for Data Table}
\usage{
mod_tbl_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for displaying a data table.
}
\description{
User Interface for Data Table
}
\section{Outputs}{

\itemize{
\item \code{year_table}: Data table.
}
}

\seealso{
\code{\link[=mod_tbl_server]{mod_tbl_server()}}
}
