% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_var_input.R
\name{mod_var_input_ui}
\alias{mod_var_input_ui}
\title{Variable Input Module - UI}
\usage{
mod_var_input_ui(id)
}
\arguments{
\item{id}{\emph{(character)} Namespace ID for the module.}
}
\value{
A \code{shiny::tagList()} containing UI elements for variable
selection.
}
\description{
Creates a UI for selecting variables for a scatter plot.
}
\section{Details}{

\code{mod_var_input_ui()} provides UI controls for customizing scatter plots.
It includes:
\itemize{
\item \strong{Dropdowns}:
\itemize{
\item X-axis variable
\item Y-axis variable
\item Color variable
}
}

The function uses \code{shiny::NS()} to namespace all input IDs.
}

\examples{

if (interactive()) {
  shiny::shinyApp(
    ui = shiny::fluidPage(mod_var_input_ui("vars")),
    server = function(input, output, session) {
      selected_vars <- mod_var_input_server("vars")
      shiny::observe(
        print(selected_vars())
      )
    }
  )
}

}
\seealso{
\itemize{
\item \code{\link[=mod_var_input_server]{mod_var_input_server()}}  for server-side logic.
\item \href{https://shiny.rstudio.com/reference/shiny/latest/tagList.html}{\code{shiny::tagList()}}
for containing UI elements.
\item \href{https://shiny.rstudio.com/reference/shiny/latest/NS.html}{\code{shiny::NS()}}
for namespacing.
}
}
\concept{\strong{Variable Input Module}}
