% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_var.R
\name{mod_var_server}
\alias{mod_var_server}
\title{var_input Server Functions}
\usage{
mod_var_server(id)
}
\arguments{
\item{id}{module namespace id}
}
\value{
shiny server module
}
\description{
var_input Server Functions
}
