% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_vbar.R
\name{mod_vbar_server}
\alias{mod_vbar_server}
\title{Server Logic for Horizontal Bar Plot}
\usage{
mod_vbar_server(id, vals)
}
\arguments{
\item{id}{Shiny module ID.}

\item{vals}{Reactive list of plot parameters.}
}
\value{
Renders a horizontal bar plot based on selected parameters.
}
\description{
Server Logic for Horizontal Bar Plot
}
\section{Reactive Inputs}{

\itemize{
\item \code{vals}: Reactive list of plot parameters.
}
}

\seealso{
\code{\link[=mod_vbar_ui]{mod_vbar_ui()}}
}
