% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_vbar.R
\name{mod_vbar_ui}
\alias{mod_vbar_ui}
\title{User Interface for Horizontal Bar Plot}
\usage{
mod_vbar_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for displaying a horizontal bar plot.
}
\description{
User Interface for Horizontal Bar Plot
}
\section{Outputs}{

\itemize{
\item \code{ratings}: Horizontal bar plot.
}
}

\seealso{
\code{\link[=mod_vbar_server]{mod_vbar_server()}}
}
