% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visualization.R
\name{mod_visualization_server}
\alias{mod_visualization_server}
\title{Visualization Module Server}
\usage{
mod_visualization_server(id, data, inputs)
}
\arguments{
\item{id}{Module namespace}

\item{data}{Reactive data}

\item{inputs}{List of reactive inputs from inputs module}
}
\value{
Reactive plot object
}
\description{
Visualization Module Server
}
