% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visualization.R
\name{mod_visualization_ui}
\alias{mod_visualization_ui}
\title{Visualization Module UI}
\usage{
mod_visualization_ui(id)
}
\arguments{
\item{id}{Module namespace}
}
\value{
Shiny UI element
}
\description{
Visualization Module UI
}
