% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthApp.R
\name{monthApp}
\alias{monthApp}
\title{Month App Package Standalone App Function}
\usage{
monthApp(...)
}
\arguments{
\item{...}{arguments passed to \code{shinyApp()}}
}
\value{
a shiny application object
}
\description{
Month App Package Standalone App Function
}
