% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthAppPkg-package.R
\docType{package}
\name{monthAppPkg-package}
\alias{monthAppPkg}
\alias{monthAppPkg-package}
\title{monthAppPkg: An example app-pkg}
\description{
A shiny application built inside an R package.
}
\keyword{internal}
