% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthFeedback.R
\name{monthFeedbackServer}
\alias{monthFeedbackServer}
\title{Month feedback (server)}
\usage{
monthFeedbackServer(id, month)
}
\arguments{
\item{id}{module ID}

\item{month}{month input}
}
\value{
text feedback on month selection
}
\description{
Month feedback (server)
}
