% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_case.R
\name{name_case}
\alias{name_case}
\title{Change the case of text}
\usage{
name_case(x, case = "title")
}
\arguments{
\item{x}{A character vector.}

\item{case}{One of \code{"title"} or \code{"lower"}. Optional.}
}
\value{
A character vector with the case changed according to the specified format.
Will error if input is not a character vector or if an unsupported case is specified.
}
\description{
Change the case of text
}
