% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar_server.R
\name{navbar_server}
\alias{navbar_server}
\title{Application Navigation Server Logic}
\usage{
navbar_server(input, output)
}
\arguments{
\item{input, output}{Shiny server input and output objects.}
}
\value{
Server logic for the Shiny application with navigation.
}
\description{
Application Navigation Server Logic
}
\section{Modules}{

\itemize{
\item \code{scatter_values}: Values for scatter plot.
\item \code{count_values}: Selected year value.
}
}

\seealso{
\code{\link[=navbar_ui]{navbar_ui()}}
}
