% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar_ui.R
\name{navbar_ui}
\alias{navbar_ui}
\title{User Interface for the Shiny Application}
\usage{
navbar_ui()
}
\value{
A tagList containing the UI elements.
}
\description{
This function defines the user interface for the Shiny application,
integrating Bootstrap and a custom JavaScript file.
}
\section{Navigation Panels}{

\itemize{
\item \code{Metrics}: Contains scatter plot settings and display.
\item \verb{By Year}: Contains year-based plots and table.
}
}

\seealso{
\code{\link[=navbar_server]{navbar_server()}}
}
