% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Inverted versions of in, is.null and is.na}
\usage{
x \%not_in\% y
}
\arguments{
\item{x}{object 1}

\item{y}{object 2}
}
\description{
Inverted versions of in, is.null and is.na
}
\examples{
1 \%not_in\% 1:10

}
