% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plot_display.R
\name{plot_display_server}
\alias{plot_display_server}
\title{plot_display Server}
\usage{
plot_display_server(id, var_inputs)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
plot_display Server
}
\keyword{internal}
