% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plot_display.R
\name{plot_displayUI}
\alias{plot_displayUI}
\title{plot_display UI}
\usage{
plot_displayUI(id)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\value{
shiny UI module
}
\description{
A shiny Module.
}
