% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_utils_server.R
\name{point_plot}
\alias{point_plot}
\title{Plot points (shiny)}
\usage{
point_plot(df, x_var, y_var, col_var, alpha_var, size_var)
}
\arguments{
\item{df}{input dataset (tibble or data.frame)}

\item{x_var}{x variable}

\item{y_var}{y variable}

\item{col_var}{color variable}

\item{alpha_var}{alpha value}

\item{size_var}{size value}
}
\value{
plot object
}
\description{
Plot points (shiny)
}
