% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint-utils.R
\name{register_endpoint}
\alias{register_endpoint}
\title{Register an Endpoint}
\usage{
register_endpoint(data, fn, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{data}{Dataset to expose/reprocess by \code{fn}.}

\item{fn}{Function to preprocess \code{data}. This function
must accept 1) the \code{data} and 2) the \code{request} from
the browser. This function must return a response,
see \link{http_response_json}.}

\item{session}{A valid shiny session object.}
}
\description{
Register an endpoint to make data accessible from the
session's browser.
}
\keyword{internal}
