% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint-utils.R
\name{register_endpoint_json}
\alias{register_endpoint_json}
\title{Register a JSON Endpoint}
\usage{
register_endpoint_json(data, fn, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{data}{Dataset to expose/reprocess by \code{fn}.}

\item{fn}{Function to preprocess \code{data}. This function
must accept 1) the \code{data} and 2) the \code{request} from
the browser. This function must return a dataset,
to serialise to JSON \link{http_response_json}.}

\item{session}{A valid shiny session object.}
}
\description{
Register an endpoint to make data accessible from the
session's browser.
}
\keyword{internal}
