% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{rvtl}
\alias{rvtl}
\title{Typing reactiveValuesToList is too long}
\usage{
rvtl(...)
}
\arguments{
\item{...}{arguments from \code{shiny::reactiveValuesToList()}}
}
\description{
Typing reactiveValuesToList is too long
}
