% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_template}
\alias{save_template}
\title{Save report template to the proper location}
\usage{
save_template(template_content)
}
\arguments{
\item{template_content}{output from \code{create_template_dir()}}
}
\value{
message regarding template creation
}
\description{
Save report template to the proper location
}
