% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plotly.R
\name{scatter_plotly}
\alias{scatter_plotly}
\title{Create a Plotly Scatter Plot}
\usage{
scatter_plotly(data, xvar, yvar, colvar)
}
\arguments{
\item{data}{A reactive expression that returns a \code{tibble} or \code{data.frame}
containing the data.}

\item{xvar}{A reactive expression specifying the column name for the x-axis
variable.}

\item{yvar}{A reactive expression specifying the column name for the y-axis
variable.}

\item{colvar}{A character string specifying the column name for the
categorical variable used for coloring points.}
}
\value{
A \code{plotly} object representing the scatter plot.
}
\description{
The \code{scatter_plotly()} function generates an interactive Plotly
scatter plot using a reactive dataset.
The user specifies the x-axis, y-axis, and a categorical variable for
coloring the points.
}
