% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stones}
\alias{stones}
\title{Birthstones data}
\format{
A data frame with 12 rows and 2 variables:
\describe{
\item{month}{name of month (January--December)}
\item{stone}{name of stone}
}
}
\usage{
stones
}
\description{
A dataset containing the month and associated birthstone. The variables are
as follows:
}
\keyword{datasets}
