% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_logo.R
\name{super_text_logo}
\alias{super_text_logo}
\title{Super text logo}
\usage{
super_text_logo()
}
\arguments{
\item{size}{header size ("h1", "h2", "h3", "h4", "h5", "h6")}
}
\value{
HTML text logo
}
\description{
Super text logo
}
\examples{
super_text_logo()
}
