% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-utils.R
\name{tag2}
\alias{tag2}
\title{Tag 2}
\usage{
tag2(.name, ...)
}
\arguments{
\item{.name}{Name of the tag.}

\item{...}{Attributes and children.}
}
\description{
Very similar to \code{\link[htmltools:builder]{htmltools::tag()}} but uses
variadic \code{...} instead of a list.
}
