% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tests.R
\name{test_logger}
\alias{test_logger}
\title{tests}
\usage{
test_logger(start = NULL, end = NULL, msg)
}
\value{
The return value, if any, from executing the utility.
}
\description{
A test utility function
}
