% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_var_input.R
\name{var_input_server}
\alias{var_input_server}
\title{var_input Server}
\usage{
var_input_server(id)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\value{
shiny server module
}
\description{
var_input Server
}
\keyword{internal}
