% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_var_input.R
\name{var_inputUI}
\alias{var_inputUI}
\title{var_input UI}
\usage{
var_inputUI(id)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\value{
shiny UI module
}
\description{
A shiny Module.
}
\keyword{internal}
