% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_nav_server.R
\name{app_nav_server}
\alias{app_nav_server}
\title{Application Navigation Server Logic}
\usage{
app_nav_server(input, output)
}
\arguments{
\item{input, output}{Shiny server input and output objects.}
}
\value{
Server logic for the Shiny application with navigation.
}
\description{
Application Navigation Server Logic
}
\section{Modules}{

\itemize{
\item \code{scatter_values}: Values for scatter plot.
\item \code{year_value}: Selected year value.
}
}

\seealso{
\code{\link[=app_nav_ui]{app_nav_ui()}}
}
