% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_nav_ui.R
\name{app_nav_ui}
\alias{app_nav_ui}
\title{User Interface for the Shiny Application}
\usage{
app_nav_ui()
}
\value{
A tagList containing the UI elements.
}
\description{
This function defines the user interface for the Shiny application,
integrating Bootstrap and a custom JavaScript file.
}
\section{Navigation Panels}{

\itemize{
\item \code{Metrics}: Contains scatter plot settings and display.
\item \verb{By Year}: Contains year-based plots and table.
}
}

\seealso{
\code{\link[=app_nav_server]{app_nav_server()}}
}
