% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_movie_awards.R
\name{create_movie_awards}
\alias{create_movie_awards}
\title{Create a Movie Awards Data Frame}
\usage{
create_movie_awards(df, award, start_year, end_year)
}
\arguments{
\item{df}{movies data.frame}

\item{award}{character string}

\item{start_year}{integer}

\item{end_year}{integer}
}
\value{
A data frame
}
\description{
Movie awards (Film or Actor).
}
\examples{
create_movie_awards(movies, "Films", 2000, 2005)
create_movie_awards(movies, "Stars", 1990, 1994)
}
