% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_movie_awards.R
\name{create_movie_awards}
\alias{create_movie_awards}
\title{Create a Filtered Movie Awards Data Frame}
\usage{
create_movie_awards(award, start_year, end_year)
}
\arguments{
\item{award}{A character string specifying the type of award to filter. Can be "pic" for Best Picture/Director or "act" for Best Actor/Actress.}

\item{start_year}{An integer specifying the starting year of the filter range.}

\item{end_year}{An integer specifying the ending year of the filter range.}
}
\value{
A data frame containing movies that won the specified award within the given year range.
}
\description{
This function filters movie awards data based on a specified award type and a given year range.
It checks that \code{start_year} is less than \code{end_year} and that the range is no greater than 5 years.
}
\examples{
create_movie_awards("pic", 2000, 2005)
create_movie_awards("act", 1990, 1994)
}
