% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_movie_awards.R
\name{create_movie_awards}
\alias{create_movie_awards}
\title{Create a Data Frame of Movie Awards}
\usage{
create_movie_awards(movies)
}
\arguments{
\item{movies}{A data frame containing movie data, including award columns
that start with "best".}
}
\value{
A data frame with columns \code{title}, \code{award}, \code{win}, and \code{thtr_rel_date}.
}
\description{
This function processes a movie data frame to create a tidy data frame of
movie awards.
}
\details{
The \code{create_movie_awards} function reshapes the input movie data frame to
extract information about various awards. It identifies columns that start
with "best", indicating different awards such as best picture, best actor,
etc. The function pivots these columns into a long format, creating a new
data frame where each row represents a movie and an associated award. The
award column names are cleaned up for readability by replacing abbreviations
with full words, and the \code{win} column values are standardized to "Yes" and
"No". The resulting data frame includes the movie title, theater release
date, award type, and whether the award was won.
}
\examples{
str(create_movie_awards(movies))

}
