% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_movie_wins.R
\name{create_movie_wins}
\alias{create_movie_wins}
\title{Convert movies data into awards wins (support function)}
\usage{
create_movie_wins(df, award, start_year, end_year)
}
\arguments{
\item{df}{movies data.frame}

\item{award}{character string ("Film" or "Stars")}

\item{start_year}{integer (start theater release year)}

\item{end_year}{integer (end theater release year)}
}
\value{
A data frame
}
\description{
Converts the movies data into the wins data (i.e., whether movies won or
lost a specified award within the given year range).
}
\examples{
create_movie_wins(movies, "Films", 2000, 2005)
create_movie_wins(movies, "Stars", 1990, 1994)

}
\seealso{
\code{\link[=filter_award_wins]{filter_award_wins()}} \code{\link[=create_movie_awards]{create_movie_awards()}}
}
