% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_award_wins.R
\name{create_movie_wins}
\alias{create_movie_wins}
\title{Create Awards Wins Data Frame}
\usage{
create_movie_wins(df, award, start_year, end_year)
}
\arguments{
\item{df}{movies data.frame}

\item{award}{A character string specifying the type of award to filter. Can be "pic" for Best Picture/Director or "act" for Best Actor/Actress.}

\item{start_year}{An integer specifying the starting year of the filter range.}

\item{end_year}{An integer specifying the ending year of the filter range.}
}
\value{
A data frame containing movies that won the specified award within the given year range.
}
\description{
Create Awards Wins Data Frame
}
\examples{
create_movie_wins(movies, "Films", 2000, 2005)
create_movie_wins(movies, "Stars", 1990, 1994)
}
