% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_award_wins.R
\name{create_movie_wins}
\alias{create_movie_wins}
\title{Create Awards Wins Data Frame}
\usage{
create_movie_wins(df, award, start_year, end_year)
}
\arguments{
\item{df}{movies data.frame}

\item{award}{character string}

\item{start_year}{integer}

\item{end_year}{integer}
}
\value{
A data frame
}
\description{
Create data containing whether movies won or lost a specified award within the given year range.
}
\examples{
create_movie_wins(movies, "Films", 2000, 2005)
create_movie_wins(movies, "Stars", 1990, 1994)

}
