% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_movies_app.R
\name{ggp2_movies_app}
\alias{ggp2_movies_app}
\title{Development \code{ggplot2movies} app standalone function}
\usage{
ggp2_movies_app(test = TRUE, run = "w")
}
\arguments{
\item{test}{logical, run in \code{test.mode}? Defaults to \code{TRUE}.}
}
\value{
shiny app
}
\description{
Wrapper function for \code{shinyAppDir()}
}
