% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Movies Review Application}
\value{
\code{ui} argument in \code{shiny::shinyApp()}
}
\description{
Starts the Movies Review Shiny application, which provides a customizable
scatter plot interface for analyzing movie data.
}
\details{
The \code{\link[=launch_app]{launch_app()}} function is as a wrapper for \code{shiny::shinyApp()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{shinyApp(ui = movies_ui, server = movies_server)
}\if{html}{\out{</div>}}

In \code{\link[=launch_app]{launch_app()}}:
\itemize{
\item UI is stored in \code{movies_ui()}
\item server is stored in \code{\link[=movies_server]{movies_server()}}
}
}
\section{\code{var_input} module}{

\code{\link[=mod_var_input_ui]{mod_var_input_ui()}} is used to collect the following inputs:
\itemize{
\item \code{input$x}
\item \code{input$y}
\item \code{input$z}
\item \code{input$alpha}
\item \code{input$size}
\item \code{input$plot_title}
}
}

\section{\code{scatter_display} module}{

\code{\link[=mod_scatter_display_ui]{mod_scatter_display_ui()}} displays the graph output using \code{\link[=scatter_plot]{scatter_plot()}}
}

\seealso{
\code{\link[=mod_var_input_server]{mod_var_input_server()}}

\code{\link[=mod_scatter_display_server]{mod_scatter_display_server()}}
}
