% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Movies Review Application}
\usage{
launch_app(app = NULL, options = list(), run = "p", ...)
}
\arguments{
\item{app}{which app to run. Options are:
\itemize{
\item \code{NULL} = the default app (\code{"movies"})
\item \code{"bslib"} = alternative \code{bslib} layout
\item \code{"ggp2"} = \code{ggplot2movies} (tidy) data app.
\item \code{"quarto"} = Quarto movies app.
}}

\item{options}{arguments to pass to \code{options()}}

\item{run}{where to launch app:
\itemize{
\item \code{p} = launch in viewer pane
\item \code{b} = launch in external browser
\item \code{w} = launch in window
}}

\item{...}{arguments passed to UI}
}
\value{
A \strong{Shiny application} object.
}
\description{
Starts the Movies Review Shiny application, which provides a customizable
scatter plot interface for analyzing movie data.
}
\details{
See the \href{https://github.com/hadley/ggplot2movies}{ggplot2movies} package.
}
\section{Details}{

The application uses:
\itemize{
\item \strong{UI}: Defined in \code{\link[=movies_ui]{movies_ui()}}.
\item \strong{Server Logic}: Defined in \code{\link[=movies_server]{movies_server()}}.
}
}

\seealso{
\itemize{
\item \code{\link[=movies_ui]{movies_ui()}} for the user interface.
\item \code{\link[=movies_server]{movies_server()}} for the server logic.
}
}
\concept{\strong{Standalone Application}}
