% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Movies Explorer Shiny Application}
\usage{
launch_app(...)
}
\arguments{
\item{...}{Additional arguments passed to shinyApp()}
}
\value{
Shiny app object
}
\description{
Launch the Movies Explorer Shiny Application
}
