% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Shiny Application}
\usage{
launch_app(options = list(), run = "p", ...)
}
\arguments{
\item{options}{A list of options for the \code{shinyApp} function.}

\item{run}{A character string specifying the display type, default is \code{"p"}.}

\item{...}{Additional arguments passed to the UI function.}
}
\description{
This function initializes and launches the Shiny application with specified
options and settings.
}
\details{
\code{launch_app()} checks if the R session is interactive. If so, it sets the
display type based on the \code{run} parameter. Then, it initializes and runs the
Shiny application using \code{shinyApp}, with the UI and server components
provided by \code{\link[=app_nav_ui]{app_nav_ui()}} and \code{\link[=app_nav_server]{app_nav_server()}}, respectively.
}
\section{Parameters}{

\itemize{
\item \code{options}: A list of options passed to the \code{shinyApp} function.
\item \code{run}: Display type for the app, default is \code{"p"}.
\item \code{...}: Additional arguments passed to the UI function.
}
}

\section{Interactive Mode}{

If the function is called in an interactive R session, it will set the
display type using \code{display_type(run = run)}.
}

\seealso{
\code{\link[=app_nav_ui]{app_nav_ui()}}, \code{\link[=app_nav_server]{app_nav_server()}}
}
