% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_awards_tbl.R
\name{mod_awards_tbl_server}
\alias{mod_awards_tbl_server}
\title{Awards Table Server Module}
\usage{
mod_awards_tbl_server(id, vals)
}
\arguments{
\item{id}{Module's ID for namespacing}

\item{vals}{Reactive providing values for award type and year range}
}
\description{
Server function for the awards table display. Takes award selection and year
range values from a reactive source to generate a table showing movie
awards data.
}
\seealso{
mod_awards_tbl_ui The corresponding UI function
}
