% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_awards_vars.R
\name{mod_awards_vars_server}
\alias{mod_awards_vars_server}
\title{Server Logic for Count Tab Input}
\usage{
mod_awards_vars_server(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A reactive list containing the selected years and award type.
}
\description{
Processes and validates user inputs for the awards section, ensuring
start and end years are properly sequenced. Provides the selected
year range and award type to other components.
}
\seealso{
\code{\link[=mod_awards_vars_ui]{mod_awards_vars_ui()}} The corresponding UI function.
}
