% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_awards_vars.R
\name{mod_awards_vars_server}
\alias{mod_awards_vars_server}
\title{Server Logic for Count Tab Input}
\usage{
mod_awards_vars_server(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A reactive list containing the selected year.
}
\description{
Server Logic for Count Tab Input
}
\section{Reactive Values}{

\itemize{
\item \code{award}: Selected award variable (\code{Films} or \code{Stars})
\item \code{start_year}: Selected 'Theater Release' start year.
\item \code{end_year}: Selected 'Theater Release' end year.
}
}

\seealso{
\code{\link[=mod_awards_vars_ui]{mod_awards_vars_ui()}}
}
