% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_awards_vars.R
\name{mod_awards_vars_ui}
\alias{mod_awards_vars_ui}
\title{User Interface for Awards Tab Input}
\usage{
mod_awards_vars_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for selecting a numeric input.
}
\description{
User Interface for Awards Tab Input
}
\section{Inputs}{

\itemize{
\item \code{start_year}: Selected 'Theater Release' start year.
\item \code{end_year}: Selected 'Theater Release' end year.
}
}

\seealso{
\code{\link[=mod_awards_vars_server]{mod_awards_vars_server()}}
}
