% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compare_point.R
\name{mod_compare_point_ui}
\alias{mod_compare_point_ui}
\title{User Interface for Scatter Plot}
\usage{
mod_compare_point_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for displaying a scatter plot.
}
\description{
User Interface for Scatter Plot
}
\section{Outputs}{

\itemize{
\item \code{scatter}: Scatter plot.
}
}

\seealso{
\code{\link[=mod_compare_point_server]{mod_compare_point_server()}}
}
