% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compare_vars.R
\name{mod_compare_vars_server}
\alias{mod_compare_vars_server}
\title{Server function for comparing variables}
\usage{
mod_compare_vars_server(id)
}
\arguments{
\item{id}{A single string identifying the module.}
}
\value{
A reactive list containing the plot parameters: x variable, y variable,
color, alpha, size, and title.
}
\description{
Server function for comparing variables
}
