% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compare_vars.R
\name{mod_compare_vars_ui}
\alias{mod_compare_vars_ui}
\title{User Interface for Variable Selection}
\usage{
mod_compare_vars_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for selecting variables and plot attributes.
}
\description{
User Interface for Variable Selection
}
