% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compare_vars.R
\name{mod_compare_vars_ui}
\alias{mod_compare_vars_ui}
\title{User Interface for Variable Selection}
\usage{
mod_compare_vars_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for selecting variables and plot attributes.
}
\description{
User Interface for Variable Selection
}
\section{Inputs}{

\itemize{
\item \code{var_x}: X variable.
\item \code{var_y}: Y variable.
\item \code{color}: Color variable.
\item \code{alpha}: Alpha transparency.
\item \code{size}: Point size.
\item \code{plot_title}: Plot title.
}
}

\seealso{
\code{\link[=mod_compare_vars_server]{mod_compare_vars_server()}}
}
