% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_counts_bar.R
\name{mod_counts_bar_ui}
\alias{mod_counts_bar_ui}
\title{User Interface for Horizontal Bar Plot}
\usage{
mod_counts_bar_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for displaying a horizontal bar plot.
}
\description{
User Interface for Horizontal Bar Plot
}
\section{Outputs}{

\itemize{
\item \code{hbar_graph}: Horizontal bar plot.
}
}

\seealso{
\code{\link[=mod_counts_bar_server]{mod_counts_bar_server()}}
}
