% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_counts_tbl.R
\name{mod_counts_tbl_server}
\alias{mod_counts_tbl_server}
\title{Server Logic for Counts Data Table}
\usage{
mod_counts_tbl_server(id, vals)
}
\arguments{
\item{id}{Shiny module ID.}

\item{vals}{Reactive list of table parameters.}
}
\value{
Renders a data table based on selected parameters.
}
\description{
Server Logic for Counts Data Table
}
\section{Reactive Inputs}{

\itemize{
\item \code{vals}: Reactive list of table parameters.
}
}

\seealso{
\code{\link[=mod_counts_tbl_ui]{mod_counts_tbl_ui()}}
}
