% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_counts_vars.R
\name{mod_counts_vars_server}
\alias{mod_counts_vars_server}
\title{Server Logic for Count Tab Input}
\usage{
mod_counts_vars_server(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A reactive list containing the selected year.
}
\description{
Server Logic for Count Tab Input
}
