% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_counts_waffle.R
\name{mod_counts_waffle_server}
\alias{mod_counts_waffle_server}
\title{Server Logic for Vertical Bar Plot}
\usage{
mod_counts_waffle_server(id, vals)
}
\arguments{
\item{id}{Shiny module ID.}

\item{vals}{Reactive list of plot parameters.}
}
\value{
Renders a vertical bar plot based on selected parameters.
}
\description{
Server Logic for Vertical Bar Plot
}
\section{Reactive Inputs}{

\itemize{
\item \code{vals}: Reactive list of plot parameters.
}
}

\seealso{
\code{\link[=mod_counts_waffle_ui]{mod_counts_waffle_ui()}}
}
