% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dist_raincloud.R
\name{mod_dist_raincloud_server}
\alias{mod_dist_raincloud_server}
\title{Server Logic for Raincloud Plot}
\usage{
mod_dist_raincloud_server(id, vals)
}
\arguments{
\item{id}{Shiny module ID.}

\item{vals}{Reactive list of plot parameters.}
}
\value{
Renders a horizontal bar plot based on selected parameters.
}
\description{
Server Logic for Raincloud Plot
}
\section{Reactive Inputs}{

\itemize{
\item \code{vals}: Reactive list of plot parameters.
}
}

\seealso{
\code{\link[=mod_dist_raincloud_ui]{mod_dist_raincloud_ui()}}
}
