% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dist_tbl.R
\name{mod_dist_tbl_ui}
\alias{mod_dist_tbl_ui}
\title{User Interface for Distribution Data Table}
\usage{
mod_dist_tbl_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for displaying a data table.
}
\description{
User Interface for Distribution Data Table
}
\section{Outputs}{

\itemize{
\item \code{dist_table}: Data table.
}
}

\seealso{
\code{\link[=mod_dist_tbl_server]{mod_dist_tbl_server()}}
}
