% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dist_vars.R
\name{mod_dist_vars_server}
\alias{mod_dist_vars_server}
\title{Server Logic for Variable Selection}
\usage{
mod_dist_vars_server(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A reactive list of selected variables and plot attributes.
}
\description{
Server Logic for Variable Selection
}
\section{Reactive Values}{

\itemize{
\item \code{num_var}: numeric variable.
\item \code{chr_var}: group variable.
\item \code{alpha}: Alpha transparency.
\item \code{size}: Point size.
}
}

\seealso{
\code{\link[=mod_dist_vars_ui]{mod_dist_vars_ui()}}
}
