% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dist_vars.R
\name{mod_dist_vars_ui}
\alias{mod_dist_vars_ui}
\title{User Interface for Distribution Variable Selection}
\usage{
mod_dist_vars_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
A UI for selecting variables and plot attributes.
}
\description{
User Interface for Distribution Variable Selection
}
\section{Inputs}{

\itemize{
\item \code{num_var}: numeric variable.
\item \code{chr_var}: group variable.
\item \code{alpha}: Alpha transparency.
\item \code{size}: Point size.
}
}

\seealso{
\code{\link[=mod_dist_vars_server]{mod_dist_vars_server()}}
}
