% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_list_to_df.R
\name{mod_list_to_df}
\alias{mod_list_to_df}
\title{Convert a Named List into a Data Frame}
\usage{
mod_list_to_df(lst)
}
\arguments{
\item{lst}{A named list where names are in the format \code{"namespace-id"}, and
values  can be character, integer, or numeric vectors.}
}
\value{
A \code{data.frame} with three columns:
The first half of the data frame contains actual values, while the second half contains data types.
\itemize{
\item \code{ns} - The namespace extracted from the name (e.g., \code{"selects"}, \code{"sliders"}, \code{"text"}).
\item \code{id} - The identifier extracted from the name (e.g., \code{"genre"}, \code{"xvar"}, \code{"reviews"}).
\item \code{value} - The associated value(s) as a character string, or the corresponding data type (\code{"chr"}, \code{"int"}, \code{"num"}).
}
}
\description{
This function takes a named list, where names follow the format
\code{"namespace-id"}, and converts it into a structured data frame with columns
for namespace, id, values, and their respective data types.
}
