% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_report.R
\name{mod_report_download_server}
\alias{mod_report_download_server}
\title{Download report module (Server)}
\usage{
mod_report_download_server(
  id,
  compare_values,
  count_values,
  dist_values,
  award_values
)
}
\arguments{
\item{id}{module id}

\item{compare_values}{reactive values from compare tab}

\item{count_values}{reactive values from counts tab}

\item{dist_values}{reactive values from distribution tab}

\item{award_values}{reactive values from awards tab}
}
\value{
downloaded report
}
\description{
Download report module (Server)
}
