% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report_download.R
\name{mod_report_download_ui}
\alias{mod_report_download_ui}
\title{Download R Markdown Report (UI)}
\usage{
mod_report_download_ui(id)
}
\arguments{
\item{id}{}
}
\value{
download button for report
}
\description{
Download R Markdown Report (UI)
}
