% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter_display.R
\name{mod_scatter_display_server}
\alias{mod_scatter_display_server}
\title{Scatter Plot Display Module - Server}
\usage{
mod_scatter_display_server(id, var_inputs, aes_inputs)
}
\arguments{
\item{id}{\emph{(character)} Namespace ID for the module.}

\item{var_inputs}{\emph{(reactive)} A reactive expression containing
user-selected variables.}

\item{aes_inputs}{\emph{(reactive)} A reactive expression containing
user-selected attributes.}
}
\value{
No direct return value. This function generates a plot output.
}
\description{
Handles the server-side logic for rendering a scatter plot.
}
\section{Details}{

\code{mod_scatter_display_server()}:
\itemize{
\item Uses \code{var_inputs} to dynamically generate a scatter plot with
user-selected variables.
\item Reads from the \code{movies} dataset, which must be loaded in the app
environment.
\item Processes plot titles and axis labels to improve readability.
}
}

\section{Reactive Inputs}{

\itemize{
\item \code{var_inputs()$x}: X-axis variable.
\item \code{var_inputs()$y}: Y-axis variable.
\item \code{var_inputs()$z}: Color aesthetic variable.
\item \code{aes_inputs()$alpha}: Transparency level.
\item \code{aes_inputs()$size}: Size of points.
\item \code{aes_inputs()$plot_title}: Title of the plot.
}
}

\examples{
if (interactive()) {
  shiny::shinyApp(
    ui = shiny::fluidPage(
      mod_var_input_ui("vars"),
      mod_aes_input_ui("aes"),
      mod_scatter_display_ui("plot")
    ),
    server = function(input, output, session) {
      selected_vars <- mod_var_input_server("vars")
      selected_aes <- mod_var_input_server("aes")
      mod_scatter_display_server("plot", selected_vars, selected_aes)
    }
  )
}

}
\seealso{
\itemize{
\item \code{\link[=mod_var_input_server]{mod_var_input_server()}} for variable selection.
\item \code{\link[=mod_aes_input_server]{mod_aes_input_server()}} for aesthetics selection.
\item \code{\link[=scatter_plot]{scatter_plot()}} for generating the scatter plot.
}

Other \strong{Plot Display Module}:
\code{\link[=mod_scatter_display_ui]{mod_scatter_display_ui()}}
}
\concept{\strong{Plot Display Module}}
