% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter_display.R
\name{mod_scatter_display_ui}
\alias{mod_scatter_display_ui}
\title{Scatter Plot Display Module - UI}
\usage{
mod_scatter_display_ui(id)
}
\arguments{
\item{id}{\emph{(character)} Namespace ID for the module.}
}
\value{
A \code{shiny::tagList} containing the plot output and metadata.
}
\description{
Creates a UI for displaying a scatter plot with user-selected options.
}
\section{Details}{

\code{mod_scatter_display_ui()} includes:
\itemize{
\item A \strong{scatter plot} created dynamically based on user input.
\item Metadata about the app, including the data source.
}
}

\seealso{
\itemize{
\item \code{\link[=mod_scatter_display_server]{mod_scatter_display_server()}}  for the server-side logic of this
module.
\item \code{\link[=scatter_plot]{scatter_plot()}}  for the utility function generating the plot.
}

Other \strong{Plot Display Module}: 
\code{\link{mod_scatter_display_server}()}
}
\concept{\strong{Plot Display Module}}
