% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter_display.R
\name{mod_scatter_display_ui}
\alias{mod_scatter_display_ui}
\title{Scatter plot display module (UI)}
\usage{
mod_scatter_display_ui(id)
}
\arguments{
\item{id}{UI module id}
}
\value{
module UI (HTML)
}
\description{
Scatter plot display module (UI)
}
\seealso{
Other {"scatter plot module functions"}: 
\code{\link{mod_scatter_display_server}()}
}
\concept{{"scatter plot module functions"}}
\keyword{internal}
