% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_slider_inputs.R
\name{mod_slider_input_server}
\alias{mod_slider_input_server}
\title{slider inputs shiny module (server)}
\usage{
mod_slider_input_server(id)
}
\arguments{
\item{id}{shiny module id}
}
\description{
slider inputs shiny module (server)
}
