% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_text_inputs.R
\name{mod_text_input_server}
\alias{mod_text_input_server}
\title{text input shiny module (server)}
\usage{
mod_text_input_server(id)
}
\arguments{
\item{id}{shiny module id}
}
\description{
text input shiny module (server)
}
