% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_var_input.R
\name{mod_var_input_ui}
\alias{mod_var_input_ui}
\title{Variable input module (UI)}
\usage{
mod_var_input_ui(id)
}
\arguments{
\item{id}{UI module id}
}
\value{
module UI (HTML)
}
\description{
\code{var_input} collects the following graph inputs:
\itemize{
\item \code{input$x}
\item \code{input$y}
\item \code{input$z}
\item \code{input$alpha}
\item \code{input$size}
\item \code{input$plot_title}
}
}
\seealso{
\code{\link[=mod_var_input_server]{mod_var_input_server()}}
}
\keyword{internal}
