% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movies}
\alias{movies}
\title{IMDB movies data}
\format{
A data frame with 651 rows and 34 variables:
\describe{
\item{title}{movie title}
\item{title_type}{type, fct (Documentary, Feature Film, TV Movie)}
\item{genre}{movie genre, fct (Action & Adventure, Animation, etc.}
\item{runtime}{movie length in minutes, num, avg = 106, sd = 19.4}
\item{mpaa_rating}{movie rating, fct (G, NC-17, PG, PG-13, R, Unrated)}
\item{studio}{name of studio, chr}
\item{thtr_rel_date}{Theatre release date, POSIXct, min = 1970-05-19 21:00:00, max = 2014-12-24 21:00:00}
\item{thtr_rel_year}{Theatre release year, num, min = 1970, max = 2014}
\item{thtr_rel_month}{Theatre release month, num, min = 1, max =12}
\item{thtr_rel_day}{Theatre release day, num, min = 1, max =31}
\item{dvd_rel_date}{DVD release date, POSIXct, min = 1991-03-27 21:00:00, max = 2015-03-02 21:00:00}
\item{dvd_rel_year}{DVD release year, num, min = 1991, max = 2015}
\item{dvd_rel_month}{DVD release month, num, min = 1, max = 12}
\item{dvd_rel_day}{DVD release day, num, min = 1, max = 31}
\item{imdb_rating}{Internet movie database rating, avg = 6.49, sd = 1.08}
\item{imdb_num_votes}{Internet movie database votes, avg = 57533, sd = 112124}
\item{critics_rating}{Rotten tomatoes rating, fct (Certified Fresh, Fresh, Rotten)}
\item{critics_score}{Rotten tomatoes score, avg = 57.7, sd = 28.4}
\item{audience_rating}{Audience rating, fct (Spilled, Upright)}
\item{audience_score}{Audience score, avg = 62.4, sd = 20.2}
\item{best_pic_nom}{Best picture nomination, fct (no, yes)}
\item{best_pic_win}{Best picture win, fct (no, yes)}
\item{best_actor_win}{Best actor win, fct (no, yes)}
\item{best_actress_win}{Best actress win, fct (no, yes)}
\item{best_dir_win}{Best director win, fct (no, yes)}
\item{top200_box}{Top 20 box-office, fct (no, yes)}
\item{director}{Name of director, chr}
\item{actor1}{Name of leading actor, chr}
\item{actor2}{Name of supporting actor, chr}
\item{actor3}{Name of #3 actor, chr}
\item{actor4}{Name of #4 actor, chr}
\item{actor5}{Name of #5 actor, chr}
\item{imdb_url}{IMDB URL}
\item{rt_url}{Rotten tomatoes URL}
}
}
\usage{
movies
}
\description{
Movie review data. Note: these data come from the \href{https://rstudio-education.github.io/shiny-course/}{Building Web Applications with shiny course}.
}
\details{
Read more about acquiring these data in the \href{https://datasciencebox.org/02-exploring-data#web-scraping-and-programming}{'Web Scraping and programming' section of Data science in a box}
}
\keyword{datasets}
