% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_app.R
\name{movies_app}
\alias{movies_app}
\title{Movies app standalone function}
\usage{
movies_app(options = list(), run = "p")
}
\arguments{
\item{options}{arguments to pass to \code{options()}}

\item{run}{where to launch app:
\itemize{
\item \code{p} = launch in viewer pane
\item \code{b} = launch in external browser
\item \code{w} = launch in window
}}

\item{...}{arguments passed to UI}
}
\value{
shiny app
}
\description{
Wrapper function for \code{shinyApp()}
}
\seealso{
\code{\link[=mod_var_input_ui]{mod_var_input_ui()}}, \code{\link[=mod_var_input_server]{mod_var_input_server()}}, \code{\link[=mod_scatter_display_ui]{mod_scatter_display_ui()}}, \code{\link[=mod_scatter_display_server]{mod_scatter_display_server()}}
}
