% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_server.R
\name{movies_server}
\alias{movies_server}
\title{Server Logic for the Movies Review Application}
\usage{
movies_server(input, output, session)
}
\arguments{
\item{input}{The Shiny `input` object.}

\item{output}{The Shiny `output` object.}

\item{session}{The Shiny `session` object.}
}
\value{
No direct return value. Defines server logic for reactive outputs 
and module integration.
}
\description{
Handles the server-side logic for the Movies Review application.
}
\section{Details}{

The server integrates the following modules:
- **Variable Input Module**: Processes user input for scatter plot 
  customization using  [`mod_var_input_server()`].
- **Scatter Plot Display Module**: Generates and displays a scatter plot 
  based on user inputs using [`mod_scatter_display_server()`].
}

\examples{
if (interactive()) {
  shiny::shinyApp(ui = movies_ui(), server = movies_server)
}

}
\seealso{
- [`movies_ui()`] for the corresponding user interface.
- [`launch_app()`] for the entry point of the application.

Other **Application Components**:
\code{\link[=movies_ui]{movies_ui()}}
}
\concept{**Application Components**}
