% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_server.R
\name{movies_server}
\alias{movies_server}
\title{Movies server function}
\value{
\code{server} argument in \code{shiny::shinyApp()}
}
\description{
Server function for standalone app function
}
\details{
The \code{\link[=launch_app]{launch_app()}} function is as a wrapper for \code{shiny::shinyApp()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{shinyApp(movies_ui, movies_server)
}\if{html}{\out{</div>}}

In \code{\link[=launch_app]{launch_app()}}:
\itemize{
\item UI is stored in \code{\link[=movies_ui]{movies_ui()}}
\item server is stored in \code{movies_server()}
}
}
\section{\code{var_input} module}{

\code{\link[=mod_var_input_server]{mod_var_input_server()}} returns following reactive values:
\itemize{
\item \code{x}
\item \code{y}
\item \code{z}
\item \code{alpha}
\item \code{size}
\item \code{plot_title}
}
}

\section{\code{scatter_display} module}{

\code{\link[=mod_scatter_display_server]{mod_scatter_display_server()}} displays the \code{ggplot2} graph with the \code{\link[=scatter_plot]{scatter_plot()}} function.
}

\section{Communication}{

The output from \code{\link[=mod_var_input_server]{mod_var_input_server()}} should be supplied to the
\code{var_inputs} argument of \code{\link[=mod_scatter_display_server]{mod_scatter_display_server()}}.
}

\seealso{
\code{\link[=mod_var_input_ui]{mod_var_input_ui()}}

\code{\link[=mod_scatter_display_ui]{mod_scatter_display_ui()}}
}
