% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_ui.R
\name{movies_ui}
\alias{movies_ui}
\title{User Interface for the Movies Review Application}
\usage{
movies_ui(bslib = FALSE)
}
\arguments{
\item{bslib}{View bslib logo?}
}
\value{
A Shiny \code{tagList} object containing the UI elements.
}
\description{
Creates the user interface (UI) for the Movies Review application, which
allows users to create customizable scatter plots based on movie data.
}
\section{Details}{

The interface is built using \href{https://rstudio.github.io/bslib/}{\code{bslib}}
\itemize{
\item \strong{Page (fillable)}: \href{https://rstudio.github.io/bslib/reference/page_fillable.html}{\code{bslib::page_fillable()}}
displays the app title.
\item \strong{Sidebar}: \href{https://rstudio.github.io/bslib/reference/sidebar.html}{\code{bslib::layout_sidebar()}}
includes a logo and the variable
selection module.
(\code{\link{mod_var_input_ui}}).
\item \strong{Card}: \href{https://rstudio.github.io/bslib/reference/card.html}{\code{bslib::card()}}
displays the scatter plot module
(\code{\link{mod_scatter_display_ui}}).
}
}

\seealso{
\itemize{
\item \code{\link[=movies_server]{movies_server()}} for the server logic of the app.
\item \code{\link[=mod_var_input_ui]{mod_var_input_ui()}} and \code{\link[=mod_scatter_display_ui]{mod_scatter_display_ui()}} for the modules
included in the UI.
}

Other \strong{Application Components}: 
\code{\link{movies_server}()}
}
\concept{\strong{Application Components}}
