% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_ui.R
\name{movies_ui}
\alias{movies_ui}
\title{User Interface for the Movies Review Application}
\usage{
movies_ui()
}
\value{
A Shiny `tagList` object containing the UI elements.
}
\description{
Creates the user interface (UI) for the Movies Review application, which 
allows users to create customizable scatter plots based on movie data.
}
\section{Details}{

The interface is built using:
- **Theme**: `shinythemes::shinytheme("spacelab")`.
- **Title Panel**: Displays the app title and a logo.
- **Sidebar Panel**: Includes the variable selection module 
  ([`mod_var_input_ui`]).
- **Main Panel**: Displays the scatter plot module 
  ([`mod_scatter_display_ui`]).
}

\examples{
if (interactive()) {
  shiny::shinyApp(ui = movies_ui(), server = movies_server)
}

}
\seealso{
- [`movies_server()`] for the server logic of the app.
- [`mod_var_input_ui()`] and [`mod_scatter_display_ui()`] for the modules 
  included in the UI.

Other **Application Components**:
\code{\link[=movies_server]{movies_server()}}
}
\concept{**Application Components**}
