% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-utils.R
\name{offcanvas}
\alias{offcanvas}
\alias{offCanvas}
\alias{offCanvasHeader}
\alias{offCanvasBody}
\alias{offCanvasDismiss}
\alias{offCanvasToggle}
\title{OffCanvas}
\usage{
offCanvas(id, ..., class = "", .position = c("right", "left", "top", "bottom"))

offCanvasHeader(..., class = "")

offCanvasBody(..., class = "")

offCanvasDismiss()

offCanvasToggle(target, color = "primary", class = "")
}
\arguments{
\item{...}{Passed to div.}

\item{class}{Additional class attribute to assign
to parent div.}

\item{.position}{Position of the offcanvas.}
}
\description{
OffCanvas
}
