% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_display.R
\name{point_plot}
\alias{point_plot}
\title{Plot points (shiny)}
\usage{
point_plot(df, x_var, y_var, col_var, alpha_var, size_var)
}
\arguments{
\item{df}{input dataset (tibble or data.frame)}

\item{x_var}{x variable}

\item{y_var}{y variable}

\item{col_var}{color variable}

\item{alpha_var}{alpha value}

\item{size_var}{size value}
}
\value{
plot object
}
\description{
Plot points (shiny)
}
\examples{
\donttest{
load(
  list.files(
    system.file("extdata", package = "lap"),
   pattern = "movies",
   full.names = TRUE)
   )
point_plot(df = movies,
  x_var = "critics_score",
  y_var = "imdb_rating",
  col_var = "critics_rating",
  alpha_var = 1/3,
  size_var = 2)
}
}
