% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scatter_display.R
\name{scatter_plot}
\alias{scatter_plot}
\title{scatter plot utility function}
\usage{
scatter_plot(df, x_var, y_var, col_var, alpha_var, size_var)
}
\value{
The return value, if any, from executing the utility.
}
\description{
A utils function
}
