% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mod_scatter_display.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Create scatter plot}
\usage{
scatter_plot(df, x_var, y_var, col_var, alpha_var, size_var)
}
\arguments{
\item{df}{\code{data.frame} or \code{tibble}}

\item{x_var}{string variable mapped to \code{x} axis}

\item{y_var}{string variable mapped to \code{y} axis}

\item{col_var}{string variable mapped to \code{color}}

\item{alpha_var}{number for point \code{alpha}}

\item{size_var}{number for point \code{size}}
}
\value{
A \code{ggplot2} plot object
}
\description{
Custom \href{https://ggplot2.tidyverse.org/}{\code{ggplot2}} function for building scatter plots in \code{moviesApp()}.
}
\examples{
scatter_plot(mtcars, 
    x_var = "mpg", 
    y_var = "disp", 
    col_var = "cyl", 
    alpha_var = 0.5, 
    size_var = 3)
    
}
\seealso{
\code{\link[=mod_scatter_display_server]{mod_scatter_display_server()}}
}
