% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{test_logger}
\alias{test_logger}
\title{Test logger (test utility)}
\usage{
test_logger(msg, start = NULL, end = NULL)
}
\arguments{
\item{msg}{test message}

\item{start}{test start message}

\item{end}{test end message}
}
\value{
message to test output
}
\description{
Test logger (test utility)
}
