% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_logger.R
\name{test_logger}
\alias{test_logger}
\title{Test logger (test utility)}
\usage{
test_logger(start = NULL, end = NULL, msg)
}
\arguments{
\item{start}{test start message}

\item{end}{test end message}

\item{msg}{test message}
}
\value{
message to test output
}
\description{
Test logger (test utility)
}
