% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_logo.R
\name{text_logo}
\alias{text_logo}
\title{Text logo}
\usage{
text_logo(size = "h3")
}
\arguments{
\item{size}{header size ("h1", "h2", "h3", "h4", "h5", "h6")}
}
\value{
HTML text logo
}
\description{
Text logo
}
\examples{
text_logo()
}
