% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_logo.R
\name{text_logo}
\alias{text_logo}
\title{Create a Text-based Logo in HTML}
\usage{
text_logo(size = "h3")
}
\arguments{
\item{size}{A character string specifying the heading size for the logo.
Valid options include "h1", "h2", "h3", "h4", "h5", and "h6".
Defaults to "h3".}
}
\value{
An HTML object containing the logo. This object can be directly used
in R Markdown documents or Shiny applications to render the logo in a web page.
}
\description{
\code{text_logo()} generates a text-based logo enclosed within HTML tags, allowing
for the specification of the heading size (h1 through h6). The logo features
a stylized representation of a face within arrows, using ASCII characters.
}
\examples{
# Generate a logo with default size h3
text_logo()

# Generate a logo with size h1
text_logo(size = "h1")


}
